Moved attribute functions to `attributes.py`
