+++ idelib/attributes.py
